@@ -4,7 +4,7 @@ namespace Rocket.Surgery.Conventions.Configuration.Yaml;
 
 internal class StaticConfigurationProvider : ConfigurationProvider
 {
-    public StaticConfigurationProvider(IDictionary<string, string> data)
+    public StaticConfigurationProvider(IDictionary<string, string?> data)
     {
         Data = data ?? throw new ArgumentNullException(nameof(data));
     }
