@@ -2,10 +2,12 @@ using Microsoft.Extensions.Configuration;
 
 namespace Rocket.Surgery.Conventions.Configuration.Yaml;
 
-internal class StaticConfigurationSource: IConfigurationSource
+internal class StaticConfigurationSource : IConfigurationSource
 {
-    public IDictionary<string,string> Data { get; set; }
+    public IDictionary<string, string?> Data { get; set; } = new Dictionary<string, string?>();
 
     public IConfigurationProvider Build(IConfigurationBuilder builder)
-        => new StaticConfigurationProvider(Data);
+    {
+        return new StaticConfigurationProvider(Data);
+    }
 }
