@@ -5,12 +5,17 @@ using YamlDotNet.Core;
 namespace Rocket.Surgery.Conventions.Configuration.Yaml;
 
 /// <summary>
-/// A YAML file based <see cref="FileConfigurationProvider"/>.
+///     A YAML file based <see cref="FileConfigurationProvider" />.
 /// </summary>
 public class YamlConfigurationProvider : FileConfigurationProvider
 {
+    /// <summary>
+    ///     The yaml configuration provider
+    /// </summary>
+    /// <param name="source"></param>
     public YamlConfigurationProvider(YamlConfigurationSource source) : base(source) { }
 
+    /// <inheritdoc />
     public override void Load(Stream stream)
     {
         var parser = new YamlConfigurationStreamParser();
