@@ -4,10 +4,11 @@ using Microsoft.Extensions.Configuration;
 namespace Rocket.Surgery.Conventions.Configuration.Yaml;
 
 /// <summary>
-/// A YAML file based <see cref="FileConfigurationSource"/>.
+///     A YAML file based <see cref="FileConfigurationSource" />.
 /// </summary>
 public class YamlConfigurationSource : FileConfigurationSource
 {
+    /// <inheritdoc />
     public override IConfigurationProvider Build(IConfigurationBuilder builder)
     {
         EnsureDefaults(builder);
