@@ -3,6 +3,8 @@ using Microsoft.Extensions.DependencyModel;
 using Microsoft.Extensions.Hosting;
 using Rocket.Surgery.Conventions;
 
+#pragma warning disable CS8620 // Argument cannot be used for parameter due to differences in the nullability of reference types.
+
 // ReSharper disable once CheckNamespace
 namespace Rocket.Surgery.Hosting;
 
@@ -19,10 +21,10 @@ public static class RocketBooster
     public static Func<IHostBuilder, ConventionContextBuilder> ForDependencyContext(DependencyContext dependencyContext)
     {
         return builder =>
-        {
-            var contextBuilder = new ConventionContextBuilder(builder.Properties).UseDependencyContext(dependencyContext);
-            return RocketHostExtensions.Configure(builder, contextBuilder);
-        };
+               {
+                   var contextBuilder = new ConventionContextBuilder(builder.Properties).UseDependencyContext(dependencyContext);
+                   return RocketHostExtensions.Configure(builder, contextBuilder);
+               };
     }
 
     /// <summary>
@@ -73,13 +75,13 @@ public static class RocketBooster
     )
     {
         return builder =>
-        {
-            var contextBuilder = new ConventionContextBuilder(builder.Properties)
-                                 // ReSharper disable once NullableWarningSuppressionIsUsed RedundantSuppressNullableWarningExpression
-                                .UseDependencyContext(DependencyContext.Default!)
-                                .WithConventionsFrom(conventionProvider);
-            return RocketHostExtensions.Configure(builder, contextBuilder);
-        };
+               {
+                   var contextBuilder = new ConventionContextBuilder(builder.Properties)
+                                        // ReSharper disable once NullableWarningSuppressionIsUsed RedundantSuppressNullableWarningExpression
+                                       .UseDependencyContext(DependencyContext.Default!)
+                                       .WithConventionsFrom(conventionProvider);
+                   return RocketHostExtensions.Configure(builder, contextBuilder);
+               };
     }
 
     /// <summary>
@@ -100,10 +102,10 @@ public static class RocketBooster
     public static Func<IHostBuilder, ConventionContextBuilder> ForAppDomain(AppDomain appDomain)
     {
         return builder =>
-        {
-            var contextBuilder = new ConventionContextBuilder(builder.Properties).UseAppDomain(appDomain);
-            return RocketHostExtensions.Configure(builder, contextBuilder);
-        };
+               {
+                   var contextBuilder = new ConventionContextBuilder(builder.Properties).UseAppDomain(appDomain);
+                   return RocketHostExtensions.Configure(builder, contextBuilder);
+               };
     }
 
     /// <summary>
@@ -137,7 +139,8 @@ public static class RocketBooster
     /// <param name="getConventions">The generated method that contains all the referenced conventions</param>
     /// <returns>Func&lt;IHostBuilder, ConventionContextBuilder&gt;.</returns>
     public static Func<IHostBuilder, ConventionContextBuilder> For(
-        AppDomain appDomain, Func<IServiceProvider, IEnumerable<IConventionWithDependencies>> getConventions
+        AppDomain appDomain,
+        Func<IServiceProvider, IEnumerable<IConventionWithDependencies>> getConventions
     )
     {
         return ForAppDomain(appDomain, getConventions);
@@ -151,10 +154,10 @@ public static class RocketBooster
     public static Func<IHostBuilder, ConventionContextBuilder> ForAssemblies(IEnumerable<Assembly> assemblies)
     {
         return builder =>
-        {
-            var contextBuilder = new ConventionContextBuilder(builder.Properties).UseAssemblies(assemblies);
-            return RocketHostExtensions.Configure(builder, contextBuilder);
-        };
+               {
+                   var contextBuilder = new ConventionContextBuilder(builder.Properties).UseAssemblies(assemblies);
+                   return RocketHostExtensions.Configure(builder, contextBuilder);
+               };
     }
 
     /// <summary>
