@@ -6,6 +6,171 @@ namespace Rocket.Surgery.Conventions.Analyzers.Tests;
 public class ImportConventionsGenericTests(ITestOutputHelper testOutputHelper) : GeneratorTest(testOutputHelper)
 {
     [Fact]
+    public async Task Should_Generate_Static_Assembly_Level_Method()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+[assembly: ImportConventions]
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+    [Fact]
+    public async Task Should_Not_Generate_Static_Assembly_Level_Method_By_Default()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+    [Fact]
+    public async Task Should_Generate_Static_Assembly_Level_Method_Custom_Namespace()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+[assembly: ImportConventions(Namespace = ""Test.My.Namespace"", ClassName = ""MyImports"")]
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+
+    [Fact]
+    public async Task Should_Generate_Static_Assembly_Level_Method_No_Namespace()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+[assembly: ImportConventions(Namespace = """", ClassName = ""MyImports"")]
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+
+    [Fact]
+    public async Task Should_Generate_Static_Assembly_Level_Method_Custom_MethodName()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+[assembly: ImportConventions(Namespace = ""Test.My.Namespace"", ClassName = ""MyImports"", MethodName = ""ImportConventions"")]
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+    [Fact]
+    public async Task Should_Generate_Static_Assembly_Level_Method_FullName()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+[assembly: ImportConventionsAttribute]
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+    [Fact]
+    public async Task Should_Generate_Static_Class_Member_Level_Method()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+namespace TestProject
+{
+    [ImportConventions]
+    public partial class Program
+    {
+    }
+}
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+
+    [Fact]
+    public async Task Should_Generate_Static_Class_Member_Level_Method_FullName()
+    {
+        var result = await WithGenericSharedDeps()
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+namespace TestProject
+{
+    [ImportConventionsAttribute]
+    public partial class Program
+    {
+    }
+}
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+    [Fact]
+    public async Task Should_Support_No_Exported_Convention_Assemblies()
+    {
+        var result = await Builder
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+
+[assembly: ImportConventions]
+"
+                           )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+    [Fact]
     public async Task Should_Support_Imports_And_Exports_In_The_Same_Assembly()
     {
         var result = await Builder
@@ -14,7 +179,7 @@ public class ImportConventionsGenericTests(ITestOutputHelper testOutputHelper) :
 using Rocket.Surgery.Conventions;
 using Rocket.Surgery.Conventions.Tests;
 
-[assembly: Convention<Contrib>]
+[assembly: Convention(typeof(Contrib))]
 
 namespace Rocket.Surgery.Conventions.Tests
 {
@@ -33,6 +198,40 @@ namespace TestProject
 }
 "
                            )
+                          .Build()
+                          .GenerateAsync();
+
+        await Verify(result);
+    }
+
+    [Fact]
+    public async Task Should_Support_Imports_And_Exports_In_The_Same_Assembly_If_Not_Exported()
+    {
+        var result = await Builder
+                          .AddSources(
+                               @"
+using Rocket.Surgery.Conventions;
+using Rocket.Surgery.Conventions.Tests;
+
+namespace Rocket.Surgery.Conventions.Tests
+{
+    [ExportConvention]
+    internal class Contrib : IConvention { }
+}
+",
+                               @"
+using Rocket.Surgery.Conventions;
+
+namespace TestProject
+{
+    [ImportConventions]
+    public partial class Program
+    {
+    }
+}
+"
+                           )
+                          .AddGlobalOption("build_property.ExportConventionsAssembly", "false")
                           .Build()
                           .GenerateAsync();
 
